+++ sstable.h
@@ -14,7 +14,7 @@ class SStable{
 private:
     // Data part
     std::string path;           // Path of the sstable file
-    uint32_t fileSize;          // Size of the sstable file
+    uint64_t fileSize;          // Size of the sstable file
 
     // Consruction part
     SSTheader * header = NULL;
@@ -29,7 +29,7 @@ public:
     // Init a SStable from a list
     SStable(uint64_t setTimeStamp, 
         std::list <std::pair<uint64_t, std::string> > &list,
-        std::string setPath, uint32_t curvLogOffset);
+        std::string setPath, uint64_t curvLogOffset);
 
     // Clear all the data
     void clear();
@@ -41,8 +41,8 @@ public:
     uint64_t getSStableKeyValNum();
 
     uint64_t getSStableKey(size_t index);
-    uint32_t getSStableKeyOffset(size_t index);
-    uint32_t getKeyIndexByKey(uint64_t key);
+    uint64_t getSStableKeyOffset(size_t index);
+    uint64_t getKeyIndexByKey(uint64_t key);
     // std::string getSStableValue(size_t index);
 
     bool checkIfKeyExist(uint64_t targetKey);
