+++ sstheader.h
@@ -13,14 +13,14 @@ public:
 
     // Constructor
     SSTheader(){};
-    SSTheader(std::string path, uint32_t offset);
+    SSTheader(std::string path, uint64_t offset);
 
     // Destructor
     ~SSTheader(){};
 
     // Load the SST header from a file
-    int readFile(std::string path, uint32_t offset);
+    int readFile(std::string path, uint64_t offset);
 
     // Save the SST header to a file
-    uint32_t writeToFile(std::string path, uint32_t offset);
+    uint64_t writeToFile(std::string path, uint64_t offset);
 };
